Updated the example with the new template git package.
