--- flutter_cache_manager/example/lib/plugin_example/cache_manager_page.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_cache_manager/flutter_cache_manager.dart';
-
-import 'download_page.dart';
-import 'floating_action_button.dart';
-
-const url = 'https://blurha.sh/assets/images/img1.jpg';
-
-/// Example [Widget] showing the functionalities of flutter_cache_manager
-class CacheManagerPage extends StatefulWidget {
-  @override
-  _CacheManagerPageState createState() => _CacheManagerPageState();
-}
-
-class _CacheManagerPageState extends State<CacheManagerPage> {
-  Stream<FileResponse> fileStream;
-
-  void _downloadFile() {
-    setState(() {
-      fileStream = DefaultCacheManager().getFileStream(url, withProgress: true);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (fileStream == null) {
-      return Scaffold(
-        appBar: null,
-        body: const ListTile(
-            title: Text('Tap the floating action button to download.')),
-        floatingActionButton: Fab(
-          downloadFile: _downloadFile,
-        ),
-      );
-    }
-    return DownloadPage(
-      fileStream: fileStream,
-      downloadFile: _downloadFile,
-      clearCache: _clearCache,
-    );
-  }
-
-  void _clearCache() {
-    DefaultCacheManager().emptyCache();
-    setState(() {
-      fileStream = null;
-    });
-  }
-}
