+++ flutter_cache_manager/example/lib/plugin_example/floating_action_button.dart
@@ -10,7 +10,7 @@ class Fab extends StatelessWidget {
     return FloatingActionButton(
       onPressed: downloadFile,
       tooltip: 'Download',
-      child: Icon(Icons.cloud_download),
+      child: const Icon(Icons.cloud_download),
     );
   }
 }
