+++ flutter_cache_manager/example/pubspec.yaml
@@ -14,6 +14,10 @@ dependencies:
     path: ../
   cupertino_icons: ^0.1.2
   url_launcher: ^5.4.11
+  baseflow_plugin_template:
+    git:
+      url: git://github.com/Baseflow/baseflow_plugin_template.git
+      ref: v1.0.0
 
 dev_dependencies:
   flutter_test:
@@ -25,5 +29,5 @@ flutter:
   uses-material-design: true
 
   assets:
-    - res/images/baseflow_logo_def_light-02.png
-    - res/images/poweredByBaseflowLogoLight@3x.png
+    - packages/baseflow_plugin_template/logo.png
+    - packages/baseflow_plugin_template/poweredByBaseflow.png
